+++ src/app/locales/en/components/content-atm-inventories-lambdas/atm-lambda-form.js
@@ -163,6 +163,9 @@ export default {
         entryIsArray: {
           label: 'Array',
         },
+        entryIsViaFile: {
+          label: 'Via file',
+        },
       },
     },
     resources: Object.assign({}, taskResourcesFields, {